refactor(configuration): move persisted agent config to objects folder
- Move `PersistedAgentConfiguration` to `Configuration/Objects`
- This improves project structure and organization.

Commit generated by The Saturn Pipeline
